hyper-v: Added 14 NET test cases
This adds the following test cases:
--NET-FILE-COPY-SCP
--NET-IP-INJECTION
--NET-TCP-CORRUPTION
--NET-PRIVATE
--NET-CHANGE-PRIVATE-NIC
--NET-JUMBO-FRAMES
--NET-COPY-LARGE-FILE
--NET-COPY-FILE-DIFFERENT-MTU
--NET-COPY-BOTH-WAYS
--NET-VLAN-TAGGING
--NET-VLAN-TRUNKING
--NET-VXLAN
--NET-MULTICAST
--ETHTOOL-CHECK-STATISTICS

This also introduces a change in the framework that will give the
possibility to run a dependency VM when needed. The framework can now
handle the switch from a setup type to another. Before this, on Hyper-V,
all the tests were performed on the setup type of the first test in
queue.