hyper-v: Cache vhds if the source path is UNC
If the OSVhd path is a local file, resume to the current workflow.

If the OSVhd path is an UNC path (ex. \smb_server\smb_share), copy the
VHD to the Hyper-V vhdx store and use that path for the parent vhd (if
the VHD is not there already).

The unicity of the VHD will be computed according to the VHD disk
identifier which will be appended to the VHD cached path.