customKernelInstall.sh: Added proposed-azure and proposed-edge
'proposed-azure' and 'proposed-edge' kernels for Ubuntu releases added to customKernelInstlal.sh
Also refreshed 'proposed' option in customKernelInstall.sh and now it accepts any Ubuntu release, not only Xenial and Trusty.
Fixed an issue with StopHyperVGroupVMs - VMs couldn't be stopped because -ComputerName param was given twice (in Hyperv.psm1 once on L560 and second on L565). That was making the function to fail and customKernelInstall for Hyper-V to fail as well.