framework: $OsVHD/$ARMImageName parameter bugfix
This is fixing an issue on Azure where lisav2 would fail if -OsVHD param was given. The framework is working with either OsVHD or ARMImageName param. But in Framework-Azure.psm1 no check was made to see what parameter is available and what parameter isn't. This is adding an extra check before manipulating each of these params.