hyperv: snapshot and turn off VM if result is not PASSED
In order to improve resource consumption on lisav2 HYPERV,
in case of a failed or aborted test, we need to create a checkpoint
and stop the VM so that it will not consume RAM/CPU on the test
environment.

This is necessary as the lisav2 hyperv test vms are created once a
new setup type is found and there are almost 200 test cases with at
least 10 setup types, and if some of the test cases are not passed,
the next VMs will not be able to start mainly due to out of RAM
exception.