performance: Hyper-V support for Network tests (#432)
Added Hyper-V support for Network Performance tests:
--This includes 2 new setupscripts and small redesign of existing
powershell scripts.
--Fixed various typos in the powershell scripts.
--Changed VM naming on Azure: the VMs will now have random names instead
of hardcoded client-vm and server-vm names. This is necessary for
Hyper-V to avoid overlapping VMs in case a previous run didn't cleanup
the VM. From now on, -role-0 vm will be considered the client-vm and
-role-1 vm will be considered the server-vm
If you have any comments on the above Azure change, please let me know. Functionality was not changed, just the VM naming.