perf_framework: iperf3: output perf results in a generic way
To compare the results for the performance testing, we need a framework
that:

is generic - should work for all performance results
has machine readable output
For genericity, all performance test cases should have a JSON output,
with the name

TestCaseName_perf_results.json

The JSON file should contain an array of objects, with the object having
the following structure
{
"Id": ValueID,
"ResultType1": Value1,
"ResultType2": Value2,
"ResultType3": Value3
}

Is mandatory to have at least the Id field.

This has to be implemented for all performance test cases to be able to
have a comparison of the results.
If the test does not implement, there will be no result comparison
possible.

The changes in the other test cases should require refactoring and
adding this feature.
This feature is backwards compatible with the previous workflow.