utilities: Added PS script for Hyper-V VM cleanup
Lisav2HyperVCleanup.ps1 will be used for Hyper-V VM cleanup. The current
cleanup process will be performed this way:
- If the VM is running for more than 48 hours, it will be stopped
- If the VM was created 5 days or earlier, it will be removed (VM files
and VHD files)
The script will also look for empty VM groups.

To run it, you need to call it with a parameter of the hostname that
will be cleaned.
e.g. .\Lisav2HyperVCleanup -HvServer "server_to_be_cleaned"