framework: Wrong InternalIP in $AllVMData
The framework didn't put InternalIP correctly. In case we had multiple
VMs (over 10), then the 'imatch' would match VMs with double digits
(e.g. client-10, client-11,) with the single digits one (e.g. with
client-1). That is breaking AllVMData since the Private IPs that are put
there are not matching the VMs.