RDMA IB: Various fixes and improvements
This commit adds a series of improvements and fixes:
- Added dynamic RHEL/CentOS Mellanox driver installation. It will
automatically detect and download the necessary drivers.
- Changed eth check from 'ifconfig' to 'ip' to avoid using net-tools
package.
- Added 'memlock unlimited' config for all distros
- Fixed Intel MPI1/RMA/NBC code. It was trying to use a variable that
doesn't exist ($modified_slaves instead of $slaves)
- All test VMs will have updated hostnames. Also, the main vm will be
updated with all the test clients in /etc/hosts.
- Resolve-UninitializedIB function will be run on all distros except
Ubuntu. It will decrease the number of random fails where port is down
and a simple reboot solves it.
- Various fixes for missing packages or files.