SR-IOV: XML refresh
The test case names and order was refreshed. New TC names are:
SRIOV-VERIFY-LSPCI
SRIOV-VERIFY-SINGLE-VF-CONNECTION
SRIOV-VERIFY-SINGLE-VF-CONNECTION-MAX-VCPU
SRIOV-VERIFY-MAX-VF-CONNECTION
SRIOV-VERIFY-MAX-VF-CONNECTION-MAX-VCPU
SRIOV-ADD-MAX-NICS-DURING-PROVISION
SRIOV-ADD-MAX-NICS-AFTER-PROVISION
SRIOV-DISABLE-ENABLE-AN
SRIOV-INTERRUPTS-CHANGE
SRIOV-DISABLEVF-ON-GUEST
SRIOV-RELOAD-MODULE
SRIOV-IPERF-STRESS

During this refresh, the script "verify-ifup-nics.sh" was removed and
it's features were integrated in NET-Max-NICs.sh. By doing this, we'll
be using the same bash script for Azure and Hyper-V testing Max NICs
test cases.