xfstests: Added CIFS Azure Files TC
This covers testing over a file share using xfstests. Azure Files
service is used for the share backend. In this new TC, a storage
account is created and then the Azure Files capability is enabled.
xfstests is started using this network share.
Also fixed a issue with btrf run (where btrfs was not installed).