dpdk: added primary secondary processes test case
DPDK primary secondary test case required:

DPDK 19.05 or greater
Linux kernel 4.20 or greater
libibverbs version 23 or greater
Currently, libibverbs need to be compiled separately, as there is no
support in Azure to install them directly.

The test case somtimes fails because the client (secondary) process
can crash because of a segmentation fault. This failure will be
fixed on the DPDK master tree.