Fix bug in process.py
While the documentation recommends that the `subprocess` module is used
via the `run()` interface, we are currently using the underlying `Popen`
interface. The documentation further recommends that if
`shell=True` (which is currently the case in our code), that the command
is passed directly as a string, not a split argument sequence. Using
this recommendation fixed the existing bug where the arguments were
being ignored on Linux (causing `uname` to print “Linux” and not the
kernel version, release, and hardware info). However, we need to revisit
our wrapping of `subprocess` because it is difficult to get right.