Improve SetSpace for serialization
The original collection of SetSpace cannot be serialized, so that azure
location cache cannot save features to disk. So add items field for
serialization.

Note, it's a minimum patch to support our functionality, not target to
create a search_space works any place else.

1. add __post_init__ to load data from deserialized data.
2. unfortunately, #1 doesn't work automatically like document. So add
decode_set_space to make it can be called explictly.
3. override add, update methods, because they are used in code.
4. call decoders in schema.