Use Makefile to centralize build/test commands
This Makefile is optional because the commands can be run independently,
it was added just to simplify and centralize all our commands. On Linux,
GNU Make is expected to almost universally be installed, but on Windows
it may need to be installed manually to use it. Fortunately the GitHub
Actions Windows image has it.