Use asyncio correctly
At minimum this converts the usage to the documented approach where this
is exactly one asyncio event loop (started in main) and everything that
currently is written async actually becomes a coroutine.

The unit tests which use coroutines now use `IsolatedAsyncioTestCase`
instead of just `TestCase`, which ensures they continue to work.