Clean up `load_runbook`
Instead of aliasing this function, we should just rename it entirely.

Instead of delegating the parsing of an `argparse.Namespace` object, we
just setup the `runbook` argument to be a `Path` instead of a string,
and have `load_runbook` accept what it needs instead of digging it out
of the `Namespace`.