Delete unnecessary `Action` class
This was an abstract class that provided no value, just extra
abstraction. Originally it was going to be used by multiple runner
implementations, and the notifier, and for artifacts. However, we only
have one runner which no longer uses it, the notifier never used it, and
artifacts aren’t implemented so it’s too early to say if it’s useful
there. Hence we should delete this to reduce complexity.