Remove useless `Runner` class
These functions did not need a class, they can run be themselves.
There’s some work to fix the functions which still rely entirely on
side-effects, but this reduces a lot of complexity (as evidenced by the
changes in the tests, half of which no longer generate a `Runner`
instance at all).