Add second smoke test example using multiple small tests
I had tried this previously, and at the time I liked it but it wasn’t
quite ready. This is re-done with a module scoped target fixture, and
with pytest-azurepipelines, this should display really nicely.

The only issue is that these each count as a separate test.

I’ve also modified smoke test B to use the `caplog` module so we can
capture INFO and above logs for this test, but not universally.

Since I don’t want to make the call on which is the “better” example,
I’m leaving both for people to review.