Stop matching `Target.params` when searching for targets
As we intend to have looser requirements, based on features (plus any
additional requirements comparison that gets implemented). We can no
longer assert on the params set because it contains the playbook’s
targets’ names, which mean every target would have to be unique,
preventing reuse.

Also cleanup parameterization with a loop.