Refactor to use locking context manager instead of session fixture
This uses a context manager with a file lock to safely access a shared
pool cached to disk (via Pytest’s built-in cache). Instead of having a
session-scoped `pool` fixture we actively create a `Target` instance
each time a `target` fixture is used (so this is now 1-1). This means
instances will be in a clean-state, and don’t need to persist across a
session, but also means that reusable targets must be properly cached to
disk (see the reference implementations). We handle deletion of targets
in the `configure` and `unconfigure` hooks, which are safe to call
regardless of the use of pytest-xdist, and don’t require a session.