notifier send message sequentially
After threading enabled in LISA, there may have chance to send messages
in parallel. It may break some notifier, and it's not necessary. So use
a queue to save messages, and try to keep it's ordered as possible. And
then send messages one by one.

Notice, the send progress is not async yet. It can be improved later.