poetry: update mypy to version "^0.812"
This version has more power on type assertions, and we can get rid of
a bunch of "# type: ignore" entries. My theory is that they point a
newer release of Typeshed (https://github.com/python/typeshed), so
that we have more type annotations available.

Real mistakes have been caught with this update, BTW.