Add the notion of a non-SUT node
We might encounter environments/platforms in which helper nodes are
needed. For example, think of a test suite in which interaction with a
TFTP server is part of the game. It does not mean that that server is
one of the SUT ones. We want to be able to leverage the tooling around
nodes, nevertheless, for that one too, but skipping all the test flow
for it, thus hiding it from node listing functions. Also, we do not
want init() on those nodes to create any working directory.

No support for that flag on "requirement" nodes. If that is shown to
make sense in the future, we can add that.

Finally, we now honor the yml/json's node entry "name", which was
there but only for internal setting.