🐛 Follow up to recent changes on to channel.recv_exit_status() lines
On e831ee037cdda04b11069b95bafddfe71b63ca44 we have removed all traces
of initial buffering of stdout in that function. Let's see if this
middle ground, in which we advance the file pointer by one byte (and
thus pull in some cache/extra meat to be consumed on stdout), but only
one byte, is enough not to hang on those crazy, different scenarios.

This fixes pristine Windows boxes detection again.