🚧 RFC only, not to merge
While I've just begun to investigate some weird Linux boxes, that hang
on shell.py's "try_connnect"--they absolutely should not, but them
we're talking about ancient, weird systems, like 3.x kernel--we must
guarantee that that function serves its purpose, which is *to detect
Windows boxes*.

When I arrived on LISA, it was broken and I made it work (just to find
that these weird cases now appear, but anyhow). One internal platform
of ours totally depends on talking to Windows boxes via SSH and, well,
the nodes need to be recognized as such for things to work.

I'd like, in parallel with any fix that will come with this
investigation, to have a unit test for such functionality. My initial
thoughts were to take advantage of the Windows instance on our CI, the
OpenSSH server feature/capability of Windows 10, and have a remote
note pointing to self (localhost) in a test, then to assert that it
was marked non-POSIX.

I'd reach that by enabling the feature, editing configs to allow for
key logins, enabling/starting the service, generating keys and using
those as credentials to the node, along with user, of course. The
test, finally, would only run on Windows (preferably detecting the
ones with the service running).

Problem is that OpenSSH's situation on Windows is a clusterf*^Wvery
poorly handled, and key logins generally only work with luck.

Anyone with a better idea on how to test that or with better abilities
with Windows + OpenSSH to help me with this?

Thanks!