🐛 [shell] One more try on working well with both POSIX/Windows worlds
Have the stdout channel for the just issued "cmd" command wait on
readiness, so at least Windows boxes get to read something (if they
are able to respond fast enough, but the timeout is very
generous). Linux ones will possibly (seen on older systems) never be
ready with that command, but at least we don't hang trying to read
something.