Implement Transformer
A transformer generates a couple of variables. User can define a couple
of transformers to prepare test data sequentially. For example, the
first transformer can install kernel package to an existing environment,
and then the second transformer generate a testable VHD. Well defined
transformers can be organized with different sequence. For example, the
second transformer, which is in previous example, can be used to copy a
image from kernel built too.

The subclasses of Transformer must implement two method,
1. _internal_run. It implements the actual logic of a transformer.
2. output_names. It tells which variables will be generated. It's used
   to validate variables before the real run. It lets failure happenning
   early.