legacy runner: downgrade some asserts to error log
The legacy LISAv2 has log written conflicts some times. If it happens,
the log cannot be writtern in file. If it happens on critical log lines,
which are used by LISA for test result status tracking, it causes the
assertions failures. The assertions fails the whole run currently, so
it's  hard to get the final test results. It's not worth to fail the
whole run, so change it to error level log.