runner: fix case skipped by lack of environment
The original design is to plan needed environments for test cases. It
doesn't create the environment candidate, if it looks not necessary. But
as practice, it means to plan it twice. One is to merge equaled
environments, the second is in the real run. It's hard to have
consistent algorithm. So remove the calculation on merging. Each test
result has a candidate, the extra environment won't be deployed.