New Test: check_floppy_module
The check_floppy_module is designed to ensure that the floppy module is not enabled on images running on Azure. This is accomplished by using the modprobe tool to check if the module is loaded. An image will fail the test if it is detected that the floppy module is loaded.