lisa_runner: attach fail prepare env to right test
The current way is to find a matched test result to failed preparation.
But a small req test result may be selected. For example, a simple test
result may match to a GPU env failure. It makes the simple test is
skipped.
With this fix, the env will save the source test, and attach preparation
failure to it.