result: skip test, if os doesn't match.
If one of OS mismatches, mark the test case is skipped. It assumes no
more env can meet the requirements, instead of checking the rest envs
one by one. The reason is this checking is a dynamic checking, and it
needs to be checked in each deployed environment. It may cause to deploy
a lot of environment for checking. In another hand, the OS should be the
same for all environments in the same lisa runner. So it's safe to skip
a test case on first os mismatched.