Add serial console support to QEMU orchestrator. (#1707)
* Add serial console support to QEMU orchestrator.

1. Add support for libvirt event callbacks. Libvirt requires
setting up a single process-wide event loop for handling the
events. This change provisions a separate Python thread for
this purpose.

2. Create a reader for the libvirt console stream that writes
the console output to file. (Ideally we would use the inbuilt
libvirt support for writing the console to file. But SELinux
doesn't like this for whatever reason.) This mimics the Azure
API that provides the serial console output as a log that can
be downloaded.

3. Implement the orchestrator serial console feature. This
just reads and dumps the console log file.

4. Fix a minor bug in NodeSpace schema class, that causes
a crash for some runbook configurations.

* Report node capabilities