util: fix path generator to take low order part of microseconds
'%f', on strftime(), reads as "Microsecond as a decimal number,
zero-padded to 6 digits". If we ever want to be minimally safer
against collisions on consecutive runs, let's stick to the low order
part of that, not the higher, which has a high chance of clashing,
under the same second. Still not crash proof, but much better.