Ensure QEMU console log shuts down.
In test automation, I am seeing a problem where LISA gets stuck during test shutdown.
This seems like it might be occuring during the QEMU VM teardown. However, there
isn't enough information in the logs to be able to pinpoint the exact cause. So, this
change adds extra debug logs.

In addition, I noticed there was a potential bug where the QEMU VM console log might
not close down correctly if the `recv` function hits an error. So, this change fixes
this problem.