QEMU: Add watchdog over VM delete.
There seems to be a rare issue where the process deadlocks while trying to delete
a QEMU VM. This issue is rare enough that getting a repo on a dev box seems to be
very difficult.

This change adds a 60 second watchdog timer over the VM delete operation. Usually,
a VM delete takes less than 1 second. So, this should be plenty to avoid false
positives. When the watchdog fires, all the Python threads' callstacks are printed
and then the process is hard stopped. This will hopefully provide insight into
why this issue is occurring.