DPDK: don't assume testpmd binary name
Fix the assumption that testpmd will be named dpdk-testpmd or testpmd
depending on the version. Some versions of rhel package newer
versions with the name 'testpmd' even though it's not the default
name when compiled.

This should somewhat future proof against another rename since
now we can check the build directory with the same function
in the source install and not assume either name.