Skip tests for unimplemented distros
Skip Cloud Hypervisor, libvirt, KVM tests for distributions other than
Ubuntu and CBLMariner. Lots of failures in other distros leading to
exceedingly high failure rate. Other distros can be enabled after fixing
these issues. Just to clarify: it is the implementation of these test
cases in LISA that doesn't currently support other distros. It is not
a limitation of the tests themselves.

Signed-off-by: Anirudh Rayabharam <anrayabh@microsoft.com>