Change timeout of fio process
Test case 'verify_cpu_offline_storage_workload' checks if fio process
exists after cpu hotplug.
The test fails when the number of cores in SKU is high since
the fio process exits before the cpu hotplug operation.

This increase in time does not affect the total duration of the
test as the tests kills the process.