Some tweaks for verify_mshv_stress_vm_create
The current timeout gets easily exceeded in our pipeline. So, make it
more generous. In the default configuration, the time taken by this
test increase with increase in the number of CPUs in the target machine.

Reduce the default iterations to 15. 25 is too high and time consuming.
Higher number of iterations, if needed, can still be configured via the
runbook.

Signed-off-by: Anirudh Rayabharam <anrayabh@microsoft.com>