Add sub-test-case support to JUnit notifier.
This change allows the JUnit notifier to output sub-test-case results.
Unfortunately, the JUnit format doesn't sub-test-cases. So, the
notifier adds the sub-test-cases as peers of the test-cases.

In addition, the JUnit notifier allows a test-case to start a sub-test-
case without explictly completing it. Instead, if a new sub-test-case
starts or the test-case completes, then the previous sub-test-case is
closed out automatically. This avoids a test-case needing to add
extensive try-catch blocks to ensure the sub-test-cases are closed out.

Other supporting changes:

1. Moved `stacktrace` and `is_completed()` from `TestResultMessage`
to `TestResultMessageBase`. This allows sub-test-cases to report
exception stacks.

2. Modify `create_test_result_message` so that it includes the time
since the test-case started in every message. This is to allow the
JUnit notifier to properly report how much time each test-case and
sub-test-case took.