Enable verbose output for stress-ng
The "-v" flag for stress-ng causes it to print detailed progress
information to stdout. This is helpful for the stress-ng suite since
they are long running tests and without this flag it is hard to tell
what the tool is doing. It is also helpful for debugging hung tests as
we get to know what exactly sress-ng was doing before it hung.

Make it a configurable option so that the users of the tool can opt out
if needed.