Retry ssh process spawn when it times out
Many times LISA fails with the following message:

"The remote node is timeout on execute [....<command>....]. It may be
caused by paramiko/spur not support the shell of node"

This is an intermittent issue with no known stable repro. In the
beginning of the run LISA is able to connect to the node and execute
commands. So, it is clearly not due to paramiko/spur not supporting
the shell of the node.

As a theoretical fix, add a mechanism to retry ssh process spawn if
it fails due to timeout. Retry will make LISA resilient to any transient
issues in the underlying layer.