mshv_root_tests: use Azure temp disk for storing large files
The testcase makes a copy of the disk image for each VM it
spawns. It is easy to run of space in OS disk especially
when the default value of 30GB is used.

Use the Azure temporary disk that is automatically mounted
at /mnt to store copies of the disk image.