lisa/util/process: return correct stdout for timed out processes
If a process times out while doing wait_result, no stdout is returned in
the result. This is not correct since the process might have produced
some output before timing out.

The output of the command is already being streamed into
self._log_buffer. Use it's contents to populate the stdout field of the
result in case of timeout.

The contents of the _log_buffer also includes the lisa tag as prefix for
each line (timestamp, node/environment info, log level etc.). This is
not required and unexpected since a process result usually includes the
raw stdout of the command and so stdout parsing logic may not have accounted
for the prefix.

The _log_buffer is currently used in wait_for_output() and here too the
logging prefix for each line is not required. It might accidently match
the content in the prefix instead of the actual command stdout.

To fix, instead of using the default formatter, use a separate one that
prints just the message without any prefix.