ch_tests_tool: report run test results even when command times out
The test comamnd timing out doesn't mean that no tests have been
run. It might have run some of the tests and before getting stuck.
In this scenario the results of the tests that were run before
the command stuck should be reported.

But when the test case times out there is no chance to
extract results of the tests that were run. To address this,
define two timemout values. CMD_TIME_OUT is the time within which
the Cloud Hypervisor test command should finish and CASE_TIME_OUT is
the time within which the whole LISA testcase should finish.

Keep the value of CASE_TIME_OUT a slightly higher so that results
can be extracted from the command's stdout when it times out. This
will ensure we have accurate reporting of test cases run even when
the test command times out.