Added TDX case in verify_isolation_config, and fixed bit masking
TDX was added as a valid_cvm_type. An if case was added for it as well.

Currently for TDX /w paravisor HV_SHARED_GPA_BOUNDARY_ACTIVE and
HV_SHARED_GPA_BOUNDARY_BITS are not settled so that case is left blank for now.

Also fixed bit selection logic, now using bit masks instead.