node: be able to set encoding of command execution channel
Say the terminal and shell at the other end of the connection are
really minimal (we even resort to falling back to
spur.ssh.ShellTypes.minimal, at shell.py). That can be the case for
BMC shells, and we now have usage on LISA for reaching out to those.

If we leave the defaults on encoding for such connection channels, we
might get things like:

    Exception in thread Thread-44 (_capture_output):
    Traceback (most recent call last):
      File "<...>Python\Python310\lib\threading.py", line 1016, in _bootstrap_inner
        self.run()
      File "<...>Python\Python310\lib\threading.py", line 953, in run
        self._target(*self._args, **self._kwargs)
      File "<...>Python\Python310\lib\site-packages\spur\io.py", line 73, in _capture_output
        output = self._file_in.read(1)
      File "<...>Python\Python310\lib\codecs.py", line 504, in read
        newchars, decodedbytes = self.decode(data, self.errors)
    UnicodeDecodeError: 'utf-8' codec can't decode byte 0xfe in position 0: invalid start byte

So expose encoding to the node execute() callers, to let them
customize the encoding, if need be. In our case, 'ansi' solves our
problems.