shell: be really minimal, when 'minimal' shell support detected
Right now, the infra still assumes some level of POSIX shell
conformance, even when going 'minimal'. Spur's ssh.sh will still apply
something like

    def escape_sh(value):
        return "'" + value.replace("'", "'\\''") + "'"

to enclose every shell token to be put on the line. Well, if you're
talking to some BMC-grade shell, which is *reall* nothing like a POSIX
shell, escaping every token with a single quote is an error on your
face.

Since that is already an extreme case, lets override that aspect of
final command lines creation, not to enclose things in single quotes.

Finally, since killing processes is not supported on those either,
return on any attempt to do so (and don't throw).