shell: don't take user data as regex compilation input
Spur indeed has something like

    class CommandInitializationError(Exception):
        def __init__(self, line):
            super(type(self), self).__init__(
    """Error while initializing command. The most likely cause is an unsupported shell. Try using a minimal shell type when calling 'spawn' or 'run'.
    (Failed to parse line '{0}' as integer)""".format(line)
            )

and LISA has something to try and remove these lines from logs, as in:

    _spawn_initialization_error_pattern = re.compile(
        r"(Failed to parse line \'b[\'\"](?P<linux_profile_error>.*?)[\'\"]\' as integer)"
    )

All fine, but when replacing, people were compiling the match as new
regexp! Let's fix that and search for literal matches to replace with
empty string. In the minimal shells we've been testing with, '*'
characters might come out of syntax error responses, as indeed

    *** Unknown syntax

so this was blowing on our faces with

    re.error: nothing to repeat at position 0