util: introduce a special exception for SSH spawn process timeout
For transient failures with the following message:

"The remote node is timeout on execute [...]. It may be caused by
paramiko/spur not support the shell of node."

When we are sure that the shell of the node is supported, the only
thing to do in this case is to retry. But since it raises a
LisaException it is impossible to differentiate it from other errors.

So, introduce a new exception type SshSpawnTimeoutException so that
test cases and/or tools can catch this exception and retry the command
(if they wish to).

Signed-off-by: Anirudh Rayabharam <anrayabh@linux.microsoft.com>