shell: fix mkdir with parents for Windows nodes
spurplus is unable to create parent directories when the target node is
a Windows node. This is because it expects only posix paths and its
parent creation logic doesn't work well with Windows paths. In these
cases spurplus throws OSError.

To fix this, catch the OSError exception and fallback to using the mkdir
command to create the directory.

Signed-off-by: Anirudh Rayabharam <anrayabh@microsoft.com>