sut_orchestrator: hyperv: serial console support
Configure COM ports for the VM using the Set-VMComPort cmdlet. The
serial console output from the guest would then be sent to the named
pipe specified when invoking Set-VMComPort.

The named pipe is not a regular file that can simply be read the usual
way. So, we use a helper Powershell script that uses the
NamedPipeClientStream .NET class to read the named pipe and write the
contents to a regular file. This helper script it copied to the server
node during environment deployment.

The powershell script is run a scheduled task. This way it runs in the
background and is not dependent on a Powershell session. This seems to
the most reliable way to run a Powershell script in the background. The
scheduled task is "started" during node deployment. The script exits
automatically if the pipe is no longer available (for e.g., VM deleted).
So, in the node delete path, just unregister the scheduled task.

Signed-off-by: Anirudh Rayabharam <anrayabh@microsoft.com>