ntttcp: allow set rmem to 1MiB
A request from Haiyang, add perf test with rmem set to a constant value.
rmem determines the amount of memory allocated for receiving packets.
The default can vary between distros and releases.

This commit adds a test perf_ntttcp_tcp_sriov_rmem_mib which sets the
value to 1 mibibyte explicitly. This will help detect differences betweenthe default and this constant value in perf results. It will also
help resolve confusion when checking the perf results between distros.

It's helpful to have an apples-to-apples comparison.

The commit enables this change for TCP+SRIOV only, adding other
tests may follow if it proves useful.

Note that UDP tests already set this value to a very large value.
There's an inconsistency between the tcp and udp tests and a possible
environment poisoning if those values are not reset between tests.