DPDK: Add rescind check
SRIOV rescind testing is a minefield of corner cases.
Sometimes Azure returns that the AN disable request
was succesful but then rescind doesn't occur.

We have a check for this but it results in an assertion.
Add a check that the functionality works before starting
the rescind tests and skip if this sku, cluster, or whatever
doesn't allow rescinding.