Account the new kernel config for kexec
A new kernel config, CONFIG_KEXEC_FILE is introduced in recent kernel
version. This config introduces a safer systemcall to load a kexec
kernel. Now kexec kernel can be loaded by both the old and the new
systemcalls. So, kdump test should check for either of KEXEC or
KEXEC_FILE config for determining the kernel's ability.