DpdkTestpmd: Refactor Installation Path
The installer path has grown to be unmaintainable.
We cover multiple cases:
-install from pkg manager
-install from source
   - install from git
   - install from tar
- (implicit) use already installed version

for multiple targets:
- debian
   - ubuntu 18.04 special casing
- suse
- fedora

and now need to allow all of these scenarios
for both DPDK and rdma-core (and the linux kernel).

SO... It's time for a refactor.

Commit makes the 'installer' concept object-oriented.

Adds a generic class 'Installer' into common.py
Adds child class SourceInstall and PackageManagerInstall.
Adds DpdkSourceInstall and DpdkPackageManagerInstall classes.
Adds DpdkGitInstall and DpdkTarInstall implementations of the generic
parent classes.